server/requirements.txt
fastapi
uvicorn
starlette
requests
minter-sdk
short_url
pycryptodomex